Alright, forget about the build/ subdirectory.
Actually using it requires making all the paths in CMakeLists.txt relative to
it, but all I was really looking for was a way to keep the build artifacts out
of the base directory.

The alternative is allowing the artifacts to pollute the base directory and
then .gitignoring them.  Fine by me!