Just finished building on Cygwin.  This required only a minor tweak.
I also documented a couple of optional dependencies -- Doxygen (which I
mentioned in a previous commit), and libglut-devel, which is needed for some
ArUco utilities.  We don't necessarily need the utilities, but I still wanted
to try them out.

Cygwin packages libglut-devel, but it does not provide PackageConfig files for
it, and my build was unable to discover the location of /usr/lib/libglut.dll.a
automatically.