Why write a camera calibration program?  OpenCV hs one already.
Its location is https://docs.opencv.org/2.4/_downloads/camera_calibration.cpp,
and I found it at
https://docs.opencv.org/2.4/doc/tutorials/calib3d/camera_calibration/camera_calibration.html.
It was easy to build.

Now, this thing requires either an XML file as input, and the XML file must in
turn provide sample input in the form of a video recording of a test image, a
live camera pointing to a test image, or a list of snapshots of test images.
So the next step is to get samples in those three forms and XML files that go
along with them.