Add XML files that actually allow you to *use* bin/calibrate with the sample images.
As far as utilities go, it's got plenty of room for improvement (an elaborate
XML file as mandatory input?  Requiring X just to generate an output X?ML
file?  Seriously?), but at least I got it to work.

I suppose the next step is to grab that checkerboard image and run calibrate
on one of our real Logitech cameras.  Fortunately, it only has to be done
once.